Restructure e2e test
- helpers are now all in the suite
- tests run within a single context, only one test namespace gets
  created
